only constraint 7, pizza order left and some issues
constraint2 : the program now only determine if the amount of wine
divisible by 3 or not, not make sure there will be a way to divide
these wine barrel into 2 part, one twice of the other.
constraint6: successfully write the checkMoves function which if you
give a correct list of move and a state of a grid will tell you is this
true or false. However, when compile with labeling, take too much time
to compute (right now it does not even finish the time to compute in my
computer for the 4*4 grid)