update on constraint 6 and pizza constraint
for constraint 6, I come up with a totally new solution, which I
strongly fill that it should work, although it doesn't at the moment.
For pizza constraint, the program querry will return a "delayed goal"
by now, but the actually result for number of each kind of pizza is
actually print out, so I am not sure if it is acceptable for professor.
I will ask him about it in class